requirements.txt
alembic==1.14.0
anyio==4.6.2.post1
certifi==2024.8.30
charset-normalizer==3.4.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
Mako==1.3.6
MarkupSafe==3.0.2
psycopg2==2.9.10
pyTelegramBotAPI==4.23.0
python-dotenv==1.0.1
python-telegram-bot==21.7
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.36
typing_extensions==4.12.2
urllib3==2.2.3
